fix: circular dependency with destinations zod schemas
